# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jikan API",
...
        "termsOfService": "https://jikan.moe/terms",
        "contact": {
            "name": "API Support (Discord)",
            "url": "http://discord.jikan.moe"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jikan-me/jikan-rest/blob/master/LICENSE"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://api.jikan.moe/v4",
            "description": "Jikan REST API Beta"
        }
    ],
    "paths": {
...
        "/people/{id}": {
            "get": {
                "tags": [
                    "people"
                ],
                "operationId": "getPersonById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns pictures related to the entry",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pictures_variants"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/people/{id}/anime": {
            "get": {
                "tags": [
                    "people"
                ],
                "operationId": "getPersonAnime",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns person's anime staff positions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/person_anime"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/people/{id}/voices": {
            "get": {
                "tags": [
                    "people"
                ],
                "operationId": "getPersonVoices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns person's voice acting roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/person_voice_acting_roles"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/people/{id}/manga": {
            "get": {
                "tags": [
                    "people"
                ],
                "operationId": "getPersonManga",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns person's published manga works",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/person_manga"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/people/{id}/pictures": {
            "get": {
                "tags": [
                    "people"
                ],
                "operationId": "getPersonPictures",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of pictures of the person",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/person_pictures"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "person_pictures": {
                "description": "Character Pictures",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/people_images"
                        }
                    }
                },
                "type": "object"
            },
...
            "anime_meta": {
                "properties": {
                    "mal_id": {
                        "description": "MyAnimeList ID",
                        "type": "integer"
                    },
                    "url": {
                        "description": "MyAnimeList URL",
                        "type": "string"
                    },
                    "images": {
                        "$ref": "#/components/schemas/anime_images"
                    },
                    "title": {
                        "description": "Entry title",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "manga_meta": {
                "properties": {
                    "mal_id": {
                        "description": "MyAnimeList ID",
                        "type": "integer"
                    },
                    "url": {
                        "description": "MyAnimeList URL",
                        "type": "string"
                    },
                    "images": {
                        "$ref": "#/components/schemas/manga_images"
                    },
                    "title": {
                        "description": "Entry title",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "character_meta": {
                "properties": {
                    "mal_id": {
                        "description": "MyAnimeList ID",
                        "type": "integer"
                    },
                    "url": {
                        "description": "MyAnimeList URL",
                        "type": "string"
                    },
                    "images": {
                        "$ref": "#/components/schemas/character_images"
                    },
                    "name": {
                        "description": "Entry name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "anime_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            },
                            "small_image_url": {
                                "nullable": true,
                                "description": "Small Image URL JPG",
                                "type": "string"
                            },
                            "large_image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "webp": {
                        "description": "Available images in WEBP",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL WEBP",
                                "type": "string"
                            },
                            "small_image_url": {
                                "nullable": true,
                                "description": "Small Image URL WEBP",
                                "type": "string"
                            },
                            "large_image_url": {
                                "nullable": true,
                                "description": "Image URL WEBP",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "manga_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            },
                            "small_image_url": {
                                "nullable": true,
                                "description": "Small Image URL JPG",
                                "type": "string"
                            },
                            "large_image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "webp": {
                        "description": "Available images in WEBP",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL WEBP",
                                "type": "string"
                            },
                            "small_image_url": {
                                "nullable": true,
                                "description": "Small Image URL WEBP",
                                "type": "string"
                            },
                            "large_image_url": {
                                "nullable": true,
                                "description": "Image URL WEBP",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "character_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            },
                            "small_image_url": {
                                "nullable": true,
                                "description": "Small Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "webp": {
                        "description": "Available images in WEBP",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL WEBP",
                                "type": "string"
                            },
                            "small_image_url": {
                                "nullable": true,
                                "description": "Small Image URL WEBP",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "people_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "common_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "person_anime": {
                "description": "Person anime staff positions",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "position": {
                                    "description": "Person's position",
                                    "type": "string"
                                },
                                "anime": {
                                    "$ref": "#/components/schemas/anime_meta"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "person_manga": {
                "description": "Person's mangaography",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "position": {
                                    "description": "Person's position",
                                    "type": "string"
                                },
                                "manga": {
                                    "$ref": "#/components/schemas/manga_meta"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "person_voice_acting_roles": {
                "description": "Person's voice acting roles",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "role": {
                                    "description": "Person's Character's role in the anime",
                                    "type": "string"
                                },
                                "anime": {
                                    "$ref": "#/components/schemas/anime_meta"
                                },
                                "character": {
                                    "$ref": "#/components/schemas/character_meta"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "pictures_variants": {
                "description": "Pictures Resource",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "images": {
                                    "$ref": "#/components/schemas/common_images"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}